Front 구조 수정
