--- liar-game-frontend/src/Home.html
@@ -1,36 +1,0 @@
-<html>
-  <body style="padding: 100px">
-    <div
-      style="
-        background-color: rgba(196, 196, 196, 1);
-        width: 100%;
-        height: 100%;
-        text-align: center;
-      "
-    >
-      <div
-        style="
-          position: absolute;
-          top: 50%;
-          left: 50%;
-          transform: translate(-50%, -50%);
-          width: 382px;
-        "
-      >
-        <h1 style="margin-bottom: 200px; font-size: 72px">LIER GAME</h1>
-        <input
-          style="
-            margin-bottom: 30px;
-            width: 100%;
-            font-size: 24px;
-            padding: 16px;
-          "
-        />
-        <br />
-        <button style="width: 100%; font-size: 24px; padding: 16px">
-          방생성
-        </button>
-      </div>
-    </div>
-  </body>
-</html>
